[sample_hevc_fei] Add timeout option for endurance testing
Timeout option set time limit during which pipeline will be
executed. Since the use of this option can produce too big
bitstream, pipeline resets output stream to the beginning even
if stream will be undecodable.
Also changed reporting of processed frames. The previous implementation
calculated a number of processed frames for encoded order, but for
display order, it calculated processed frames as a number of submitted
frames. If DrainBufferFrames failed, it would print the wrong number
for ENCODE in display order.