Avoid forcing CMAKE_VERBOSE_MAKEFILE: let user chose
CMAKE_VERBOSE_MAKEFILE causes output of the whole build command lines
to the console instead of just file name being compiled or linked. This
greatly obscures the output! As a result we have missed the following
gcc compiler warning as a minimum:
  cc1: warning: command line option ‘-Wnon-virtual-dtor’ is valid for C++/ObjC++ but not for C [enabled by default]

This change removes CMAKE_VERBOSE_MAKEFILE forced to true. User can
either specify it on a cmake confugre command line as:
  mkdir build && cd build
  cmake -DCMAKE_VERBOSE_MAKEFILE=TRUE ..
or simply force output during the build:
  make VERBOSE=1

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>