[vpp]: Fix random error out on VPP initialization due to non-configured attributes
Fixes: #236

This is misuse of vaGetConfigAttributes function. It _expects_ to get list of
attributes (i.e. types) to be queried. Instead mediasdk tried to do I don't know what.
It passed non-initialized attribute to the function and got random results. We could
hit unknown attribute and error out or we could hit known attribute and initialize to
some unpredictable behaviour.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>