Fix decoder hang
Decoder stucks when DPB is exhausted and submitted slices have
[SliceHeader :: pic_output_flag == 0]

Flow which cause the issue:
1. Submit 1st slice (IDR)
2. There is available data in bitstream, but there is no free frames
in allocator. This is because [async == 1] and for sys. mem an internal
allocator is used. Note, increasing async depth doesn't help, we just
move a little bit further.
3. [DecodeFrame] returns ERR_MORE_SURFACE -> no scheduler task is
created
4. Repeat 2-3 for a couple of following slices, until capacity of
internal allocator (DPB) is exhausted
6. Error above is handled in the following way - if we have enough
data to start 'decoding' (this means we have at least on full AU in
queue) - we create special 'virtual task' (if it was not scheduled yet).
7. The thing is we don't create this 'virtual task' because we have
frames which are 'not outputted' (SliceHeader :: pic_output_flag == 0]

Signed-off-by: Dmitry Ermilov <dmitry.ermilov@intel.com>