core: remove iterative encoding initialization
This patch removes iterative encoding intialization: try to create
HW class, if failed or feature is not supported in HW - try to create
SW class. There are few reasons to remove this:
1. In open source we don't have a SW class, thus second iteration
is actually doing nothing
2. This is assymetry from what all other components are doing

If SW fallback will be needed in the future, we will add this via
a wrapper encoder class which will take care of what exactly should
be created inside.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>