fastboot initial commit
The goal is to have a possibility of changing set of used codecs and components of
libmfx library.
In the _studio folder the "fastboot" subfolder contains cmake build scripts for
alternatives. In the builder folder the "profiles" subfolder contains cmake includes,
that are contains set of options (codecs and components switchers) to be used for this
alternative.
Current cmake build system builds all alternatives of libmfx library, that will be
placed in "_studo/fastboot" subfolder.