build: add ENABLE_ALL option
-DENABLE_ALL will enable all the dependencies and features unless user did not
explicitly switched some of them OFF, i.e. configuring in the following way
is possible:
  cmake -DENABLE_ALL=ON -DENABLE_TEXTLOG=OFF
and it will configure all targets except samples.

TODO: As of now ENABLE_ALL is not fully implemented, it will not affect all
  of the ENABLE_* options. Those options it don't affect are placed above
  ENABLE_ALL definition and require some rework and/or pending CI adoption.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>

[ASC] New descriptors integration

Integration of two new descriptors for ASC module, which
is preliminary work for implementation of fade detection
new feature in ASC. The new descriptors are 1) temporal
correlation and 2) motion compensated temporal correlation

Issue: -
Test: Manual test