encoders: return VAAPI rate control values as unit32_t
Accroding to VAAPI specification rate control modes are
defined as unit32_t. For example:
  #define VA_RC_CBR  0x00000002
  #define VA_RC_QVBR 0x00000400
In mediasdk code in all encoder we pretend that these values are
uint8_t which is wrong. This worked till we supported only those rate
controls which falls into 1st byte, but this will not work with,
for example, QVBR. This patch fixes that.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>