sample_decode: check and report errors in case of X11 DRI3 rendering
Fixes: #1396

In case of X11 DRI3 we try to directly render output surfaces via memory
sharing mechanism. For that we extract DMA buffer handle (file descriptor)
from video surfaces and create XCB pixmap from it, but this operation
might fail since rendering scanout might not be supported for some color
formats we use for media. For example, typically we decode into NV12 which
is not supported for scanouts. Hence the following might fail:
  sample_decode h264 -i input.264 -r -f 30

Before this patch user will see the following error which is not very
descriptive:
  X Error of failed request:  BadAlloc (insufficient resources for operation)
    Major opcode of failed request:  149 ()
    Minor opcode of failed request:  2
    Serial number of failed request:  19
    Current serial number in output stream:  22

After this patch user will get:
  Failed to create xcb pixmap from the NV12 surface: try another
    color format (e.g. RGB4)
  DeliverOutput return error = -6

Which hopefull will lead user trying to run the following command line
which should work assuming scanout from rgb4 is supported (which usually is):
  sample_decode h264 -i input.264 -r -f 30 -rgb4

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>