A new approach of MFX_WRN_DEVICE_BUSY handling
If MFX_WRN_DEVICE_BUSY is returned, it shall be paired with a valid syncPoint.
That syncPoint is not necessarily binded with actuall output data.
The purpose of the syncPoint is provide applications a way to make polling-free waiting.
Applications, of course, can choose either to wait or perform some different job.

Signed-off-by: Dmitry Ermilov <dmitry.ermilov@intel.com>