[MCTF] Segmentation fault bug fix related to missing output pointer
A segmentation faul was happening because the ouput pointer was not
being assigned when the previous frame was assigned as a P frame to
be denoised and the next frame happened to be and I frame which
also needed to be denoised. The second frame (I frame) was not able
to get the actual output pointer for itself and that caused the
segmentation fault.

Issue: https://github.com/Intel-Media-SDK/MediaSDK/issues/2412
Test:  Manual testing based on MDP-64761