[AVCe] Correct VA Interlaced caps reading
Previously DDI caps for interlace were confused, for VA_ENC_INTERLACED_FIELD (==2) ddi_caps.NoInterlacedField was set to 1, which means that interlaced field IS NOT supported, and vice versa, for VA_ENC_INTERLACED_NONE (==0) or VA_ATTRIB_NOT_SUPPORTED ddi_caps.NoInterlacedField was set to 0, which means that interlaced field IS supported.
Fixed this logic.